Command Handler
Changed the command handler to be able to function properly, rather than error on start-up. One small gripe is that the command and event handler is a slightly modified handler from https://github.com/ItzAshu397/Discord.js-v13-CMD-Event-Handler and the way this is written the token has to be hard coded into clientManager.js. To combat this i added it to the gitignore and attatched an example file, aptly titled `clientManager.example.js`.